Template: autocomplete differential pairs
If a user is selecting the source pin of a configuration, some logic was added to try and autocomplete differential pair's source pin as well. If the channel ends in a "+" or "-" and the opposite channel is found without a source pin selected, the custom device will try to autocomplete the differential pair source as well.